+++ tests/testthat/test-transformations.R
@@ -1,19 +1,19 @@
 test_that("hilo of transformed distributions", {
   expect_identical(
-    hilo(exp(dist_normal())),
-    exp(hilo((dist_normal())))
+    hilo(exp(dist_poisson(3))),
+    exp(hilo((dist_poisson(3))))
   )
 })
 
 test_that("chains of transformations", {
   expect_identical(
-    hilo(dist_normal()),
-    hilo(log(exp(dist_normal())))
+    hilo(dist_student_t(5)),
+    hilo(log(exp(dist_student_t(5))))
   )
 
   expect_output(
-    print(exp(dist_normal())-1),
-    "t\\(N\\(0, 1\\)\\)"
+    print(exp(dist_student_t(5))-1),
+    "t\\(t\\(5, 0, 1\\)\\)"
   )
 })
 
@@ -31,6 +31,17 @@ test_that("handling of transformation arguments", {
 
 test_that("LogNormal distributions", {
   dist <- dist_transformed(dist_normal(0, 0.5), exp, log)
+  ln_dist <- dist_lognormal(0, 0.5)
+
+  # Test exp() shortcut
+  expect_identical(
+    exp(dist_normal(0, 0.5)),
+    ln_dist
+  )
+  expect_identical(
+    log(ln_dist),
+    dist_normal(0, 0.5)
+  )
 
   # format
   expect_equal(format(dist), sprintf("t(%s)", format(dist_normal(0, 0.5))))
@@ -38,19 +49,19 @@ test_that("LogNormal distributions", {
   # quantiles
   expect_equal(
     quantile(dist, c(0.1, 0.5)),
-    list(qlnorm(c(0.1, 0.5), 0, 0.5))
+    quantile(ln_dist, c(0.1, 0.5))
   )
 
   # pdf
   expect_equal(
     density(dist, c(1, 20)),
-    list(dlnorm(c(1, 20), 0, 0.5))
+    density(ln_dist, c(1, 20))
   )
 
   # cdf
   expect_equal(
     cdf(dist, c(4, 90)),
-    list(plnorm(c(4, 90), 0, 0.5))
+    cdf(ln_dist, c(4, 90))
   )
 
   # F(Finv(a)) ~= a